requirements.txt
appdirs==1.4.4
asgiref==3.3.1
backports.zoneinfo==0.2.1
black==21.7b0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
#cryptography==3.4.6
# Django==3.1.5
django-cors-headers==3.7.0
django-crispy-forms==1.13.0
django-durationwidget==1.0.5
django-markdownx==3.0.1
django-rest-knox==4.1.0
django-storages==1.11.1
django-wsgi==1.0b1
djangorestframework==3.12.2
djongo==1.3.6
dnspython==2.1.0
google-api-core==2.0.1
google-auth==2.1.0
google-cloud-bigquery==2.31.0
google-cloud-core==2.0.0
google-cloud-storage==1.42.2
google-crc32c==1.2.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
grpcio==1.42.0
gunicorn==20.1.0
idna==3.2
importlib-metadata==4.8.3
Markdown==3.3.6
mypy-extensions==0.4.3
#mysqlclient==2.1.0
nibabel==3.2.1
numpy==1.18.0
packaging==21.3
pandas==1.1.3
pathspec==0.9.0
Pillow==8.1.0
proto-plus==1.19.8
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydicom==2.2.2
pymongo==4.0.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-http-client==3.3.2
pytz==2020.5
regex==2021.8.28
requests==2.26.0
rsa==4.7.2
sendgrid==6.8.1
six==1.10.0
sqlparse==0.2.4
starkbank-ecdsa==1.1.1
tomli==1.2.1
urllib3==1.26.6
WebOb==1.8.7
whitenoise==5.3.0
wincertstore==0.2
#zipp==3.6.0
